Toast favoritos

+++ src/app/services/favorites.service.ts
@@ -1,4 +1,5 @@
 import { Injectable } from '@angular/core';
+import { ToastController } from '@ionic/angular';
 import { Book } from '../interfaces/interfacesBookAll';
 import { Movie } from '../interfaces/interfacesMovies';
 
@@ -9,7 +10,7 @@ export class FavoritesService {
 
   favBooks: Book[] = [];
   favMovies: Movie[] = [];
-  constructor() { 
+  constructor(private toastController: ToastController) { 
     this.loadFavs();
   }
 
@@ -31,28 +32,58 @@ export class FavoritesService {
 
   addFavBook(book: Book)
   {
-    this.favBooks.unshift(book);
-    localStorage.setItem('favBooks', JSON.stringify(this.favBooks));
+    const exist = this.favBooks.find( b => b.primary_isbn10 === book.primary_isbn10);
+    
+    if(!exist)
+    {
+      this.favBooks.unshift(book);
+      localStorage.setItem('favBooks', JSON.stringify(this.favBooks));
+      this.showToast('Book added to favorites');
+    }
+    else
+    {
+      this.showToast('This book is already added');
+    }
   }
   
   addFavMovie(movie: Movie)
   {
-    this.favMovies.unshift(movie);
-    localStorage.setItem('favMovies', JSON.stringify(this.favMovies));
+    const exist = this.favMovies.find(m => m.display_title === movie.display_title);
+    
+    if(!exist)
+    {
+      this.favMovies.unshift(movie);
+      localStorage.setItem('favMovies', JSON.stringify(this.favMovies));
+      this.showToast('Movie added to favorites');
+    }
+    else
+    {
+      this.showToast('This movie is already added');
+    }
   }
 
   removeFavMovie(movie: Movie)
   {
     this.favMovies = this.favMovies.filter(m => m.display_title !== movie.display_title);
     localStorage.setItem('favMovies', JSON.stringify(this.favMovies));
+    this.showToast('Movie removed from favorites');
   }
-
+  
   removeFavBook(book: Book)
   {
     this.favBooks = this.favBooks.filter(b => b.primary_isbn10  !== book.primary_isbn10);
     localStorage.setItem('favBooks', JSON.stringify(this.favBooks));
+    this.showToast('Book removed from favorites');
   }
 
+  async showToast(message: string)
+  {
+    const toast = await this.toastController.create({
+      message: message,
+      duration: 500
+    });
+    toast.present();
+  }
 }
